deployment/compose/scripts/requirements.txt
requests
kafka-python
pika
